Create a new file: webpack.config.js. Here's how Webpack works: we create a config file that tells it which file to load, where to save the final file, and a few other things. Then... it does the rest!
But... Webpack's configuration is complex. A fully-featured setup will probably be a few hundred lines of complicate config! Heck, our Webpack tutorial was over 3 hours long! Very simply: Encore is a tool that helps generate that complex config.
Setting up webpack.config.js

Click on the documentation to find the 'First Example'. Hey! A webpack.config.js file to get us started! Copy that! Then, paste it in our file. But, I'm going to simplify and delete a few things: we'll add this stuff back later. Just keep setOutputPath(), setPublicPath() and addEntry().

And hey, check out that first line! Since this file will be executed by Node, we can require stuff! This imports the Encore object:

15 lines webpack.config.js
var Encore = require('@symfony/webpack-encore');
... lines 2 - 15

Then, at the bottom, we ask Encore to give us the final config, and we export it:

15 lines webpack.config.js
... lines 1 - 12
// export the final configuration
module.exports = Encore.getWebpackConfig();

There are only three things we need to tell Webpack: the directory where we want to save the final files - public/build - the public path to that directory - so /build since public/ is the document root - and an 'entry'. Point this to our JavaScript file: ./public/assets/js/RepLogApp.js. Change the first argument to rep_log. This tells Webpack to work its magic on RepLogApp.js. The first argument will be the name of the final file, .js - so rep_log.js.