Run 'yarn add sweetalert2 --dev'
Let's require one more outside package. Search for 'swal'. We're using a really cool library called SweetAlert to bring up the delete dialog. But... the only reason this works is that, in the template, we're including a script tag for it:

67 lines templates/lift/index.html.twig
... lines 1 - 47
{% block stylesheets %}
... lines 49 - 50
    <link rel='stylesheet' href='https://cdn.jsdelivr.net/npm/sweetalert2@7.11.0/dist/sweetalert2.min.css' />
{% endblock %}
... line 53
{% block javascripts %}
... lines 55 - 56
    <script src='https://cdn.jsdelivr.net/npm/sweetalert2@7.11.0/dist/sweetalert2.all.min.js'></script>
... lines 58 - 65
{% endblock %}

Boo! Let's refactor this to require that library properly.

If you search for this package, you'll find out that it's called sweetalert2.