Creating a Skinny 'entry' File
Next, look in the template: index.html.twig. We include the rep_log.js file... but we also have a little bit of JavaScript that is responsible for using that object and initializing it:
{% block javascripts %}
... lines 55 - 56
    <script src='{{ asset('build/rep_log.js') }}'></script>
    <script>
        {dollar_sign}(document).ready(function() {
            var {dollar_sign}wrapper = {dollar_sign}('.js-rep-log-table');
            var repLogApp = new RepLogApp({dollar_sign}wrapper);
        });
    </script>
{% endblock %}

This is.... kind of a bummer: it relies on the RepLogApp variable to be global... and that only works because, at the bottom, we're purposely creating a global variable with window.RepLogApp = RepLogApp:

214 lines public/assets/js/RepLogApp.js
... lines 1 - 212
window.RepLogApp = RepLogApp;

Also, to fully Webpackify our app, we will eventually want to remove all JavaScript from our templates. Yep, you'll just include the one JS file and... that's it!

Skinny Entries

And this brings us to an important point about organization. Usually, the entry file - so the file that we list in webpack.config.js:
Encore
... lines 4 - 9
    .addEntry('rep_log', './public/assets/js/RepLogApp.js')
... lines 11 - 12
;
 Should contain a small amount of logic that calls out to other modules. It's kind of like a controller in Symfony: it's supposed to have just a few lines of code that call out to other parts of our app.

Actually, the code in index.html.twig is a pretty good example of what I'd expect in an entry file:
 66 lines templates/lift/index.html.twig
... lines 1 - 53
{% block javascripts %}
... lines 55 - 58
    <script>
        {dollar_sign}(document).ready(function() {
            var {dollar_sign}wrapper = {dollar_sign}('.js-rep-log-table');
            var repLogApp = new RepLogApp({dollar_sign}wrapper);
        });
    </script>
{% endblock %}

Let me show you what I mean: in the js/ directory, create a new file: called rep_log.js.

Next, open webpack.config.js: let's use this as the entry file instead:
17 lines webpack.config.js
... lines 1 - 2
Encore
... lines 4 - 9
    .addEntry('rep_log', './public/assets/js/rep_log.js')
... lines 11 - 12
;
... lines 14 - 17

And since I just made a change, find your terminal and restart Encore:

yarn run encore dev --watch

Copy the code from index.html.twig, remove it, and paste it here:
10 lines public/assets/js/rep_log.js
... lines 1 - 5
{dollar_sign}(document).ready(function() {
    var {dollar_sign}wrapper = {dollar_sign}('.js-rep-log-table');
    var repLogApp = new RepLogApp({dollar_sign}wrapper);
});

Perfect!

And now that we are responsible JavaScript developers... finally... we need to require any dependencies. Oh, but first, add 'use strict'; on top - that's optional, but I like it. Now add const {dollar_sign} = require('jquery') and, to get RepLogApp, const RepLogApp = require('./RepLogApp').

I love it! Does it work? Move of it and... refresh! Bah!

    RepLogApp is not a constructor

Ooof. This is a technical way of saying:

    Hey! You're using RepLogApp like a class... but it's not!