Requiring the Layout CSS So let's do this everywhere. Open layout.js and also the base layout: base.html.twig. Look at the top: we have a few css files, the first is main.css. In layout.js, require this: ../css/main.css. As soon as we hit save, synchronize the build directory again... Yes! We have a new layout.css file! In base.html.twig, update the link tag to use this.
Yep... everything still looks fine.

Handling Images
But... wait! Something amazing just happened! Look inside main.css. Woh! We're referencing a background image: ../images/dumbell-mini.png:

 78 lines  public/assets/css/main.css
... lines 1 - 7
.mini-dumbbell {
... lines 9 - 10
    background: url('../images/dumbbell-mini.png') center center  no-repeat;
... line 12
}

That's a problem! Why? Because the final file lives in a completely different directory, so that ../ path will break!

Actually... it's not a problem! Webpack is amazing! It parses our CSS looking for any background images or fonts. When it finds one, it moves it into a build/images/ directory and rewrites the path inside the final CSS file to point there.

The point is: all we need to do is write our CSS files correctly and... well... Webpack takes care of the rest!