And... refresh! Woohoo! Wait... there's a missing image! Bah! I was lying! There is one file that still needs to be publicly accessible!
Open index.html.twig... ah! We have a good, old-fashioned img tag that references one of the images in the assets/ directory:

59 lines templates/lift/index.html.twig
... lines 1 - 2
{% block body %}
    <div class=row>
... lines 5 - 34
        <div class='col-md-5'>
            <div class='leaderboard'>
                <h2 class='text-center'>
                    <img class='dumbbell' src='{{ asset('assets/images/dumbbell.png') }}' />
... line 39
                </h2>
... lines 41 - 42
            </div>
        </div>
    </div>
{% endblock %}
... lines 47 - 59

And... whoops! It's not public anymore. My bad!

This is one of the few cases - maybe the only case - where we need to reference public images from outside a file that Webpack processes. The simple problem is that Webpack doesn't know that it needs to move this file!

And... whoops! It's not public anymore. My bad!

This is one of the few cases - maybe the only case - where we need to reference public images from outside a file that Webpack processes. The simple problem is that Webpack doesn't know that it needs to move this file!

Of course, there's an easy fix: we could just move this one file back into the public/ directory. But... that sucks: I'd rather keep all of my assets in one place.