Run 'yarn add copy-webpack-plugin --dev'
We can take advantage of a Webpack plugin that can copy the file for us. Google for copy-webpack-plugin to find its GitHub page(https://github.com/webpack-contrib/copy-webpack-plugin). Encore gives you a lot of features... but it doesn't give you everything. But... no worries! We're using Webpack under-the-hood. So if you find a Webpack plugin you want, you can totally use it!

Side note, Encore will have a copy() method soon. Then you'll be able to do this without a plugin. Yay! But, this is still a great example of how to extend Webpack beyond Encore.

Anyways, install the plugin first. Notice that they use npm.