Our layout.js file requires main.css That's cool... I guess... if you like using boring old CSS. But I want to be more Hipster, so let's use Sass instead. Well, I could use Stylus to be super hipster... and Encore does support that, but let's use something a bit more familiar.
To start, rename the file to main.scss. Now, we can use a fancier syntax for these pseudo-selectors.

OooooOoOooOooo.

Obviously, the build is failing because, in layout.js, that file path is wrong! Change it to main.scss. So... does it already work?

No! On the watch tab of our terminal, it failed when loading main.scss. Out-of-the-box, Encore cannot process Sass files. But, it tells you how to fix this! We just need to enable it inside our config & install some extra packages. Remember: Encore is full of features. But to stay light, it doesn't enable everything automatically. Instead, you are in control: enable what you need, and Encore will tell you what to do. It's kinda cool!